forecast/requirements.txt
numpy==1.23.4
streamlit==1.9.0
protobuf==3.20.1
werkzeug==2.0.3
altair==4.1.0
gunicorn==20.1.0
# mysqlclient
# SQLAlchemy
pandas
bokeh==2.4.1
matplotlib
autogluon.timeseries